Add option to specify remote payload name
Sometimes, Telerik UI will append ".tmp", or something similar, to the
(already renamed) remote payload on target. Updated the exploit script
to account for that during the normal upload/deserialize flow, and added
an option to explicitly set the remote payload name if only
deserializing.

Also refactored various things for modularity and readability. This is
an embarrassingly large commit, but such is life.